feat(VULKAN-151): one skinned shader serves both spellings of BlendIndices
XNA lets a content processor write BLENDINDICES as Byte4 or as Vector4, and a
Vulkan shader input cannot take both an integer and a float attribute. The five
skinned vertex shaders declared uvec4, so only the integer spelling bound and
the float one was refused by name -- where EasyGL accepts both.

They take vec4 now and subscript the palette with int(). Vector4 binds
natively; Byte4 binds through VK_FORMAT_R8G8B8A8_USCALED, which converts
integer components to float without normalising -- exactly what
glVertexAttribPointer(..., GL_UNSIGNED_BYTE, GL_FALSE, ...) does, and therefore
the conversion EasyGL has always relied on. A bone palette is a handful of
entries, far inside float's exact-integer range.

_USCALED is not a mandatory vertex-buffer format, so this is measured rather
than assumed. spikes/vulkan-vertex-format-spike/ asks each physical device:
RADV and llvmpipe both answer YES. The renderer queries the device it selected
and, where the format is absent, refuses by name and names the Vector4 spelling
as the way round, rather than letting vkCreateGraphicsPipelines fail at
Present.

The conversion is narrow on purpose: offered only where the declaration named
the input's own alternateFormat, so an arbitrary integer element that lines up
with a float input is still unrepresentable, and a float element at an integer
input always is.

Vulkan_VertexInputLayout is 16/16 -- E2 rewritten to the resolved contract, E3
added for the narrowness rule. Vulkan_DeclaredEffectLayout gains leg C2 and is
11/11: the same skinned quad drawn twice, asserted as equality of the two
readbacks as the row asked, plus a second assertion that the picture is the
sampled texel, because two blank frames are equal too. The bone index used is
1 and bone 0 puts the quad off screen, so a draw that ignored the attribute
would render nothing.

One shared test needed correcting, and it is not a weakened assertion:
GltfRendererPbrFallbackPolicy greps the shader source for
bb.bones[aBoneIndices.y], now bb.bones[int(aBoneIndices.y)] -- the shape
EasyGL's own row in that table has always had. It greps from disk, so the
failure was live in every renderer configuration until this fix.

-R '^Vulkan_' is 231/231 and the full ctest is 9069/9087 with no Vulkan_*
failure among the 18.

+++ modules/renderers/vulkan/src/shaders/pbr3d_skinned.vert.glsl
@@ -11,7 +11,15 @@ layout(location = 1) in vec3  aNormal;
 layout(location = 2) in vec4  aTangent;
 layout(location = 3) in vec2  aUV;
 layout(location = 4) in vec4  aBoneWeights;
-layout(location = 5) in uvec4 aBoneIndices;
+// plans/plan_vulkan.md VULKAN-151: `vec4`, not `uvec4`. XNA lets a content processor spell
+// BLENDINDICES as Byte4 or as Vector4 (plans/plan_fx.md FX-127, and CustomModelAnimation's own
+// SkinnedModelProcessor writes Vector4), and a Vulkan shader input cannot take both an integer and
+// a float attribute. Taking the indices as floats lets ONE shader serve both spellings: Vector4
+// binds natively, and Byte4 binds through VK_FORMAT_R8G8B8A8_USCALED -- integer values converted
+// to float without normalisation, which is exactly what EasyGL's own
+// glVertexAttribPointer(..., GL_UNSIGNED_BYTE, GL_FALSE, ...) does. A bone palette is a handful of
+// entries, far inside float's exact-integer range, so nothing is lost in the conversion.
+layout(location = 5) in vec4 aBoneIndices;
 #ifdef CNA_PBR_DUAL_UV
 layout(location = 6) in vec2  aUV1;
 #endif
@@ -84,10 +92,10 @@ float cnaDirectionHandedness(mat3 m) {
 
 void main() {
     float weightsPerVertex = pbr.fogColorEnabled.w; // REMED-GFX-010: alongside the fog vector
-    mat4 skinMat = bb.bones[aBoneIndices.x] * aBoneWeights.x;
-    if (weightsPerVertex >= 2.0) skinMat += bb.bones[aBoneIndices.y] * aBoneWeights.y;
-    if (weightsPerVertex >= 4.0) skinMat += bb.bones[aBoneIndices.z] * aBoneWeights.z
-                                          + bb.bones[aBoneIndices.w] * aBoneWeights.w;
+    mat4 skinMat = bb.bones[int(aBoneIndices.x)] * aBoneWeights.x;
+    if (weightsPerVertex >= 2.0) skinMat += bb.bones[int(aBoneIndices.y)] * aBoneWeights.y;
+    if (weightsPerVertex >= 4.0) skinMat += bb.bones[int(aBoneIndices.z)] * aBoneWeights.z
+                                          + bb.bones[int(aBoneIndices.w)] * aBoneWeights.w;
     vec4 skinnedPos = skinMat * vec4(aPos, 1.0);
     gl_Position = pc.mvp * skinnedPos;
     // REMED-GFX-011: matches skinned3d.vert.glsl, which does flip (the comment previously here
